+++ src/b1b/08/prog/04--iris-from-str.c
@@ -4,8 +4,8 @@
 
 typedef struct Feature
 {
-  float length;
-  float width;
+  double length;
+  double width;
 } Feature;
 
 typedef struct Iris
@@ -18,8 +18,8 @@ typedef struct Iris
 void dump_iris(Iris iris, const char *name)
 {
   printf("%s:\n", name);
-  printf("  Sepal: (%f, %f)\n", iris.sepal.length, iris.sepal.width);
-  printf("  Petal: (%f, %f)\n", iris.petal.length, iris.petal.width);
+  printf("  Sepal: (%lf, %lf)\n", iris.sepal.length, iris.sepal.width);
+  printf("  Petal: (%lf, %lf)\n", iris.petal.length, iris.petal.width);
   printf("  Class: %s\n", iris.class);
 }
 
@@ -27,6 +27,8 @@ Iris parse_str(const char *str)
 {
   char buf[256];
   strcpy(buf, str);
+
+  // NOTE: 各行の終端の `\n` で, `iris.class` に `\n` が含まれてしまうので削除
   buf[strlen(buf) - 1] = '\0';
 
   Iris iris;
